Drafting the new Neighborhoods API.
