+++ src/skshapes/linear_operators/__init__.py
@@ -1,0 +1,1 @@
+from .linear_operator import InverseParameters, LinearOperator
